Update Flashcard.java to better follow OOP standard
