+++ src/main/java/objects/flashcard/Flashcard.java
@@ -1,0 +1,20 @@
+package objects.flashcard;
+
+
+public class Flashcard {
+    private String questionText;
+    private String answerText;
+
+    public Flashcard(String questionText, String answerText) {
+        this.questionText = questionText;
+        this.answerText = answerText;
+    }
+
+    public String getQuestionText() {
+        return questionText;
+    }
+
+    public String getAnswerText() {
+        return answerText;
+    }
+}
